requirements.txt
django==4.1.7
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
freezegun==1.1.0
PyJWT==2.4.0
pytest-django==4.5.2
black==22.3.0
flake8==4.0.1
isort==5.10.1